fix: #autostart without line num
The autostart was missing when it used the `#autostart` (no line number) syntax - it should point to the next available line number, otherwise autostart didn't appear to work.

Python isn't my first (or second, or third) language, so I hope this is okay. I've tested it locally and compared to the generated files from the .txt2bas command on the Next (and in cspect) to confirm the bytes are loaded up correctly and this looks to be the right change.